// src/types/book/genre.enum.ts
export enum Genre {
  Fantasy = 'Fantasy',
  ScienceFiction = 'Science Fiction',
  Dystopian = 'Dystopian',
  Adventure = 'Adventure',
  Romance = 'Romance',
  Mystery = 'Mystery',
  Thriller = 'Thriller',
  Horror = 'Horror',
  HistoricalFiction = 'Historical Fiction',
  LiteraryFiction = 'Literary Fiction',
  Contemporary = 'Contemporary',
  YoungAdult = 'Young Adult',
  Children = 'Children’s',
  Biography = 'Biography',
  Memoir = 'Memoir',
  History = 'History',
  SelfHelp = 'Self-Help',
  Health = 'Health',
  Travel = 'Travel',
  TrueCrime = 'True Crime',
  Science = 'Science',
  Philosophy = 'Philosophy',
  Psychology = 'Psychology',
  Business = 'Business',
  Politics = 'Politics',
  Religion = 'Religion',
  Education = 'Education',
  Art = 'Art',
  Cooking = 'Cooking',
  Technology = 'Technology',
  Environment = 'Environment',
  Comics = 'Comics',
  GraphicNovel = 'Graphic Novel',
  Manga = 'Manga',
  Poetry = 'Poetry',
  Anthology = 'Anthology',
  Essay = 'Essay',
  ShortStories = 'Short Stories',
  Guide = 'Guide'
}